move TimePartition to DatasetConfig
